conditional display of air quality, fixed slider update, fixed slider location
