+++ src/components/pollutant-exposures/air-quality-slider.styles.ts
@@ -1,8 +1,10 @@
 import { createStyles, Slider } from '@material-ui/core';
+import { globalStyles } from '../../styles/globalStyles';
 
 import theme from "../../theme/theme";
 
 export const styles = createStyles({
+  ...globalStyles,
 
   slider: {
     width: "90%",
